requirements.txt
streamlit>=1.20

dimod>=0.10.11
dwave-neal>=0.8.1
